- version 0.8.9.16 - fixing install.py script - Installing Start Menu Shortcut on Windows

+++ pyradio/install.py
@@ -918,6 +918,8 @@ if __name__ == '__main__':
     if platform.system().lower().startswith('linux'):
         parser.add_argument('--python2', action='store_true',
                             help='install using python 2.')
+    else:
+        parser.add_argument('--python2', action='store_true', help=SUPPRESS)
     parser.add_argument('-R', '--uninstall', action='store_true',
                         help='uninstall PyRadio.')
 
@@ -944,33 +946,34 @@ if __name__ == '__main__':
     args = parser.parse_args()
     sys.stdout.flush()
 
-    if args.brew != "False":
-        if args.brew is None:
-            param = ' --fix-mac-path'
-        else:
-            param = ' --warning --fix-mac-path "' + args.brew + '"'
-        #### print('parameter: ' + param)
-        try:
-            from urllib.request import urlretrieve
-        except:
-            from urllib import urlretrieve
-        try:
-            r = urlretrieve('https://raw.githubusercontent.com/coderholic/pyradio/master/devel/build_install_pyradio')
-        except:
-            print('Cannot contact github...')
-            sys.exit(1)
-        if int(r[1]['content-length']) < 1000:
-            print('Cannot contact github...')
-            sys.exit(1)
-        script = r[0]
-        #### print('script:', script)
-        if exists('/Users/Max/pyradio/devel/build_install_pyradio'):
-            script = '/Users/Max/pyradio/devel/build_install_pyradio'
-        #### print('script:', script)
-        # now I can run the script
-        subprocess.call('bash -c "' + script + param + '"', shell=True)
-        os.remove(r[0])
-        sys.exit()
+    if platform.system().lower().startswith('darwin'):
+        if args.brew != "False":
+            if args.brew is None:
+                param = ' --fix-mac-path'
+            else:
+                param = ' --warning --fix-mac-path "' + args.brew + '"'
+            #### print('parameter: ' + param)
+            try:
+                from urllib.request import urlretrieve
+            except:
+                from urllib import urlretrieve
+            try:
+                r = urlretrieve('https://raw.githubusercontent.com/coderholic/pyradio/master/devel/build_install_pyradio')
+            except:
+                print('Cannot contact github...')
+                sys.exit(1)
+            if int(r[1]['content-length']) < 1000:
+                print('Cannot contact github...')
+                sys.exit(1)
+            script = r[0]
+            #### print('script:', script)
+            if exists('/Users/Max/pyradio/devel/build_install_pyradio'):
+                script = '/Users/Max/pyradio/devel/build_install_pyradio'
+            #### print('script:', script)
+            # now I can run the script
+            subprocess.call('bash -c "' + script + param + '"', shell=True)
+            os.remove(r[0])
+            sys.exit()
 
     if not PY3 and not args.python2:
         print_trying_to_install()
